pyproject.toml
[project]
name = "pyganttccpm"
version = "0.1.0"
#packages = ["src"]
description = "using Matplotlib to generate Gantt-style charts visualizing Critical Chain Project Management (CCPM) task networks."
readme = "README.md"
keywords = ["planning", "resource management", "task management", "collaboration", "visualization", "cross-platform"]
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
license-files = ["LICENSE.txt"]
authors = [
    {name = "R.N. Wolf", email = "rudi@rnwolf.net"},
]
maintainers = [
  {name = "Rüdiger Wolf", email = "rudi@rnwolf.net"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Topic :: Office/Business :: Scheduling",
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
    "Natural Language :: English",
]
dependencies = [
    "networkx>=3.0",
    "matplotlib>=3.10.1",
    "python-dateutil>=2.8",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "semver>=3.0.4",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.11",
    "mkdocstrings[python]>=0.29.1",
    "mkdocs-git-revision-date-localized-plugin>=1.4.5",
    "mkdocs-material>=9.6.11",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
#packages = ["src"]
#packages = find_packages(where="src")
# This will find and include the main package and all subpackages
packages = { find = { include = ["pyganttccpm"] } }

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in selected subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
# 5. Use single quotes in `ruff format`.
quote-style = "single"
